requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
playwright-stealth==2.0.3
anthropic==0.34.2
python-dotenv==1.0.1
httpx==0.27.2
plyer==2.1.0
twilio==9.3.3
requests==2.32.3
apscheduler==3.10.4
